backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Pydantic v3 (actually v2.x is the latest, v3 doesn't exist yet)
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
greenlet==3.1.1

# Redis and Celery
redis==5.2.0
celery==5.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
authlib==1.3.2
httpx==0.28.1

# ML Libraries
scikit-learn==1.6.0
xgboost==2.1.3
lightgbm==4.5.0
optuna==4.1.0
shap==0.46.0
mlflow==2.18.0

# Data processing
pandas==2.2.3
numpy==2.2.0
imbalanced-learn==0.12.4

# Object storage (S3-compatible for Cloudflare R2)
boto3==1.35.76

# Monitoring and observability
prometheus-client==0.21.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0

# Utilities
python-dotenv==1.0.1
reportlab==4.2.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
hypothesis==6.122.3
faker==33.1.0
